Update `DefaultSession.hsf` with more verbose window configuration:
 - Enable windows for registers, memory, locals, stack trace, watch

 - Use config variables for download modules to automatically track
   selected build configuration

Clean up and refactor some code to eliminate compiler warnings.

+++ MerpMod/GenericTests.c
@@ -73,7 +73,7 @@ void BrickBuster()
 
 #ifdef pRamHoleEnd
 unsigned int test = pRamHoleEnd;
-unsigned int end = &(pRamVariables.RamHoleEndMarker);
+unsigned int end = (unsigned int) &(pRamVariables.RamHoleEndMarker);
 if(test < end)
 Assert(0,"error in ram hole!");
 #endif
